+++ source/Validation/documents/release-notes/release-notes.md
@@ -1,9 +1,13 @@
 # Schema Validation Release notes
 
+## Version 1.0.6
+
+ - Added ReadValueAsDurationAsync() for xs:duration data types.
+
 ## Version 1.0.5
 
  - CanReadValue could incorrectly return false when an attribute was preceding the element content.
- - This only applies to the Reader API - extension methods AsXmlReaderAsync and AsXElementAsync are not affected.
+   This only applies to the Reader API - extension methods AsXmlReaderAsync and AsXElementAsync are not affected.
 
 ## Version 1.0.4